pyproject.toml
[project]
name = "cashflo-nlp-sql"
version = "0.1.0"
description = "Semantic Layer + NLP-to-SQL engine for Cashflo AP automation platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.30.0",
    "PyYAML>=6.0",
    "streamlit>=1.35.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "plotly>=5.20.0",
]

[project.scripts]
cashflo = "app:CashfloApp"

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["engine"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]
